+++ LoadPE/LPE_Inject.h
@@ -2,6 +2,6 @@
 #include <Windows.h>
 #include <tchar.h>
 // 远程线程注入
-HMODULE RmThread_Inject(DWORD pid, LPCSTR dll_path);
+HMODULE RmThread_Inject(long pid, LPCSTR dll_path);
 // 远程线程卸载注入
-BOOL RmThread_Unject(DWORD pid, HMODULE hmod);
+BOOL RmThread_Unject(long pid, HMODULE hmod);
